feat: NavigationGroup 추가
